Introduce memory reduction for Aerosols collection and fix UVFLUX diags
This commit registers into the geos-chem-classic wrapper the following
commits that were made in the src/GEOS-Chem submodule:

8e5ea253 Add modifications for memory reduction for the Aerosols collection
07e7c1f7 UVFLUX* diagnostics now work properly when specified individually

This brings in the diagnostic memory reduction to the Aerosols collection.
It also and fixes an issue caused by typos in the spelling of the mapping
objects for the UvFlux* diagnostic arrays in fast_jx_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>