Remove "make install" from quick-build.yml and build-matrix.yml
There seems to be an issue with the installation of the geos file
to the /usr/local/geos folder in the geos-chem-classic Azure DevOps
pipelines "Quick Build" and "Build Matrix".  For now, we will omit
the "make -j install" command in the .yml scripts. But this will need
to be addressed eventually.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>