Remove TIMERS from the CMakeLists.txt
In prior GEOS-Chem versions, TIMERS=y was used to set a Cpp switch.
In 12.9.0 and later versions, a switch in Input_Opt is now used to
toggle the GEOS-Chem timers on or off.  Therefore, we no longer need
the text that defines the TIMERS cpp switch in the CMakeLists.txt.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>