Now use -DRUNDIR_SIM=fullchem in Azure Dev Pipelines YAML files
For GEOS-Chem 13.0.0, we have consolidated all of the chemical
mechansims into a single mechanism called "fullchem".  But this update
has inadvertently broken the Azure Dev Pipelines CI builds.

The solution is to change -DRUNDIR_SIM=standard to -DRUNDIR_SIM=fullchem
in the cmake command in the ./ci-pipelines/build-matrix.yml and
./ci-pipelines/quick-build.yml files.  This will cause the code to be
configured and compiled properly on the Azure Dev Pipelines site.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>