RRTMG call time update: shift first call from 1:30 to first timestep
This update also removed the previously added kludge to reduce impact of
zeros in the GEOS-Chem History accumulator for RRTMG netcdf diagnostics
over the first 1.5 hrs of simulation. That update is no longer necessary
if RRTMG is called in the first timestep. With this update the RRTMG
netcdf and binary diagnostics match extremely well.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>