Now display GEOS-Chem and HEMCO precision to logs (uses updated HEMCO)
GEOS-Chem and HEMCO now display the precision at which they were
compiled with (REAL*4 or REAL*8) to their respective log files, as
well as from the output of the cmake configure stage.  This update
relies on changes to the HEMCO CMakeLists and CMakeScripts.

In addition, this commit informs the geos-chem classic wrapper about
the following changes that were committed to the src/GEOS-Chem submodule:

30c70b44 Now display whether flex precision is REAL*4 or REAL*8 to stdout

and to the src/HEMCO submodule:

c8402f0  Add HCO-Helpers.cmake file; also print precision to the HEMCO.log file

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>