Fix typos in GCHP HISTORY.rc.fullchem template (in StateMet collection)
This commit was made to correct a problem caused by typos in
the GCHP HISTORY.rc.fullchem template in the StateMet collection.
With these fixes, the GCHP benchmark simulation now runs properly.
(This fixes the "persistent run error" noted in the prior commit.)

The following updates were committed to the src/GEOS-Chem submodule:

3e5e959b Fix typos in GCHP HISTORY.rc.fullchem template

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>