Move main.F90 to Interfaces/GCClassic; Display Git status in configure
We have moved the GEOS-Chem "Classic" driver program main.F90 from
GeosCore to Interfaces/GCClassic.  This usage is now more consistent
with how we interface the GEOS-Chem science codebase to other
external modeling contexts such as NASA/GEOS GCM and GCHPctm.

We also now tell the src/CMakeLists.txt to look for main.F90 directly
with the path GEOS-Chem/Interfaces/GCClassic/main.F90, instead of
relying upon a symbolic link.

Also updated the CMakeLists in the GCClassic, GEOS-Chem, and HEMCO
folders to display the repository status during the configure state.
This is done using the get_repo_version CMake function, which is
included in GC-Helpers.cmake and HCO-Helpers.cmake.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>