GCHP rundir creation bug fix; recommit GCHP bug fix
(1) We needed to add a new case to for $sim_extra_option=none in the
script run/shared/setupConfigFiles.sh, so that the "@" characters in
the HISTORY.rc.fullchem template file will be removed.

(2) For some reason the fix noted in (1) was not correctly pushed to
the geos-chem repository.  This update now pushes the correct code change.

The GCClassic wrapper has been informed of the following updates
that were committed to the src/GEOS-Chem submodule:

dd726b48 Recommit GCHP bug fix in run/shared/setupConfigFiles.sh
097266a7 GCHP bug fix: Also remove "@" characters for standard simulations

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>