Now define Cmake cache variable EXE_FILE_NAME for executable filename
In src/CMakeLists.txt, we now use a cache variable EXE_FILE_NAME to
specify the default executable file name.  The default is gcclassic,
but in certain circumstances (e.g. integration tests) we may want
to rename the executable for specialty simulations (e.g. gcclassic.rrtmg,
gcclassic.apm, etc.).

To specify a different file name, use the -DEXE_FILE_NAME=___
option when running cmake (where ___ is the name that you want to use).

Thanks to Liam Bindle for the algorithm!

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>