Auto-assign never-stale label to certain Github issue types
We have updated the Github issue templates in geoschem/GCClassic to
add the never-stale label to feature requests and pull requests.  This
will make sure that the stalebot will not auto-mark them for deletion
after 14 days of inactivity.

Also, we have made the corresponding updates for the geoschem/GEOS-Chem
Github issue templates in this commit in the src/GEOS-Chem submodule:

69977d3f Auto-assign never-stale label to feature requests, discussions, PRs

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>