GEOS-Chem submod update: Bug fixes for the GEOS-Chem Classic dry-run
This commit informs the GCClassic superproject about the following
commits that were pushed to the Github geoschem/geos-chem repository:

378e79e0e Further fixes for the GEOS-Chem Classic dry run
b872b6d60 Restore the NOON_FILE_ROOT variable in ucx_mod.F90 as a local variable
991ca72aa Add updates for download_data.py for proper restart files (incomplete)
6cdd9076e Remove obsolete code from INIT_UCX that caused dry-runs to fail

These are fixes to the createRunDir.sh and download_data.py scripts,
as well as a fix for a bug in ucx_mod.F90 that caused dry-runs to fail.
These updates should only affect GEOS-Chem Classic simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>