GEOS-Chem submod update: Fix scale factors and MW for C3H8, C2H6 for AEIC
This commit informs the GCClassic superproject about the following
commits that were pushed to the Github geoschem/geos-chem repository:

173e82780 Merge branch 'bugfix/christophkeller/aeic_emissions_voc_scaling' into dev
ea31497eb Fix incorrect mol weights in species_database and HEMCO_Config templates

This update makes sure that the AEIC emissions are properly scaled
from atoms C to molec species.  Also, we have fixed the MW of EOH
in the species_database.yml to be 46.07, which is the latest accepted
value.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>