HEMCO submod update: GCHP bug fix, use all zeroes for restart field not filled
This commit informs the GCClassic superproject about the following
commits that were pushed to the Github geoschem/HEMCO repository:

218632a GCHP bug fix: revert to all zeros in criteria for restart field not filled

This had caused certain extension-related emissions (e.g. MEGAN) to
have very small values in GCHP.  Lizzie Lundgren has supplied a fix
for this issue, and we have now tagged this commit as 13.4.0-rc.2.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>