Add -g to default compile options; Remove -gdwarf-2 gstrict-dwarf from debug
We have added -g to the list of default compiler options for both
Intel Fortran and GNU Fortran.  This will help users to run or view the
gcclassic executable in a debugger without having to recompile with
-DCMAKE_BUILD_TYPE=Debug.

Also removed the -gdwarf-2 -gstrict-dwarf from the list of gfortran
debugging compiler options.  This forced using the DWARF-2 standard,
which has now been replaced by more modern versions.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>