Restore use of git tag in cloud-benchmarking primary key
This reverts commit d4cd5607ac90beb43131d3558e42f8e55a7d4880.

The previous commit was made to use the commit ID in the primary key name
to avoid issues when using multiple tags on the same commit (e.g. when
an rc tag is added as the same commit as an alpha tag). Commit IDs are
not as readable as tag names, however. Instead, we should manually trigger
1-month benchmarks for rc tags.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>