GEOS-Chem submod update: Close GC-Classic restart files after writing
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

4aa4fecf9 Close Restart collection files after each call to History_Netcdf_Write
ce51408e9 Move first call to History_Write outside of time loop in main.F90

This fixes an issue where the GEOS-Chem Classic restart files would not
be flushed to disk immediately.  This meant that if a GEOS-Chem run
stopped unexpectedly, the last restart file would be incomplete.
This now fixes the issue raised by @cdholmes in geoschem/geos-chem #1464.

NOTE: This fix only applies to GEOS-Chem Classic restart files, but
we should also extend it to other instantaneous diagnostic files
later on (as time allows).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>