GEOS-Chem submod updates: Set ESMF logging mode at run-time
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/GEOS-Chem repository:

2fd200295 Merge PR #1879 Add new GCHP config file for ESMF logging

This breaks the hardwired configuration for ESMF logfile output and
lets GCHP users specify ESMF logfile options in a configuration file.

This will not cause any differences for GEOS-Chem Classic simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>