GEOS-Chem submod update: Merge PR #1896 (Use singleton time dims)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

42159b1e8 Merge PR #1896 (Pass arrays w/ singleton time dim to netCDF routines)

This fixes an issue in History diagnostics that was originally described
in issue #1888 by @nicholasbalasus.  The fix involved using arrays
with singleton dimensions for time to pass data to netCDF-F90 library
routines.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>