GEOS-Chem submod update: Merge PR #1971 (Set halfpolar for global hi-res grids)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

094d3ccfb (Merge PR #1971 (Set half polar grid when creating global hi-res domains)

This fixes an issue where global 0.25x0.3125 or 0.5x0.625 global grids
were not getting set with the proper grid setting to use half-sized
polar gridboxes.  This is necessary for 12x12km IMI simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>