GEOS-Chem update: Merge PR #2712 (Build KPPSA upon request)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

60381876d  Merge PR #2712 (Only build KPP-Standalone upon request)

This PR now asks the user if they wish to build the KPP-Standalone
executable at run directory creation time.  A reminder will be
printed to configure with -DKPPSA=y, and the kpp_standalone_interface.yml
file will be copies to the run directory.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>