GEOS-Chem update: Merge PR #2521 (Bit-for-bit reproducability in GCC)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

b1676fe42 Merge PR #2521 (Bit-for-bit reproducibility in GCClassic)

This PR applies changes in GC-Classic needed to achieve bit-for-bit
reproducibility when breaking up a run in time. Please note that with
these updates out-of-the-box GC-Classic will still not have reproducibility
upon restart due to a limitation of HEMCO. HEMCO reads inputs, including
GEOS-Chem restart file variables, as single precision when most of these
fields in GEOS-Chem are double precision.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>