GEOS-Chem update: Merge PR #2967 (Hidden option, read GCC rst as REAL*8 )
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

efa625fe9 Merge PR #2967 (Add hidden option to read GC-Classic restart file as REAL8)

PR #2967 adds a hidden option to read the GC-Classic restart file locally
in the GEOS-Chem repository as REAL8. This option is only available for
simulations other than nested and Hg. It is a hidden feature, meaning
the entry for it is not included in the geoschem_config.yml file
templates used for making run directories. To tap into the feature
users need to manually add simulation menu entry read_restart_as_real8
and set it to true. This will result in reading the GC-Classic restart
file locally in GEOS-Chem as REAL8 and use those values rather than the
values read by HEMCO which are in REAL4.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>