GEOS-Chem update: Merge PR #2996 (Update GFEIv3 files)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

90799084d Merge PR #2996 (Update GFEIv3 files, from peer review)

PR #2996 fixes an issue in the original GFEIv3 files.  The files were
corrected in July 2025 and redownloaded and processed for use in GEOS-Chem.
The corrected files can now be found in ExtData/HEMCO/CH4/v2025-08/GFEIv3/.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>