GEOS-Chem update: Merge PR #3020 (Unhide direct GC-Classic rst read)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

e256ad747 Merge PR #3020 (Unhide feature to read GC-Classic rst file w/o HEMCO)

PR #3020 update adds the option to use GEOS-Chem to read the GC-Classic
restart file and therefore preserve native precision stored in the file.
This option has been present since 14.6.3 but it was not visible in the
geoschem_config.yml files as an option. The default is currently false.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>