GEOS-Chem update: Merge PR #3050 (Fix precision of area from FV3)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

3d8a03ceb (HEAD -> dev/14.7.0) Merge PR #3050 (Fix precision of GCHP import of area from FV3)

PR #3050 fixes a bug in GCHP where area was stored in the State_Grid
object as REAL*8 but was imported from FV3 advection as REAL*4. The area
computed in FV3 is REAL*8 so there is no reason to truncate to REAL*4
for the import. (Requires companion PR geoschem/FVdycoreCubed_GridComp#10.)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>