GEOS-Chem update: Merge PR #3074 (Fix error in carbon sim)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

89336ce92  Merge PR #3074 (Fix typos causing carbon simulations to crash)

PR #3074 fixes an out-of-bounds error in KPP/carbon/carbon_Funcs.F90
when the Carbon diagnostic collection is enabled.  The Carbon
collection is now also turned on by default so as to allow our
integration tests to catch similar errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>