GEOS-Chem update: Merge PR #3087 (Fix bug in ObsPack sampling)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

65ae1f305 Merge PR #3087 (Fix bug in ObsPack sampling)

PR #3087 updates the logic so that ObsPack data points that are
instantaneously-sampled (i.e. having CT_sampling_strategy = 4) and
that lie within 1/2 model timestep of the end of the day will be
included in the sampling.  This fixes a bug in #3084 that was
reported by @mattloman.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>