HEMCO update: Merge PR #346 (Use gcc 15 in "Mac" GitHub action)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/HEMCO repository:

HEMCO update: Merge PR #346 (Use gcc 15 in "Mac" GitHub action)

This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/HEMCO repository:

43ad195 Merge PR #346 (Replace gcc 12 w/ gcc 15 in Mac GitHub Action )

PR #346 removes GNU Compiler Collection 12 and replaces it with version 15
in the "Mac" GitHub Action, which tries to build HEMCO on MacOS.  GCC 12
evidently has been "delisted" from the Mac "HomeBrew" package manager,
and this causes the compilation test with GCC 12 to fail.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>
43ad195 Merge PR #346 (Replace gcc 12 w/ gcc 15 in Mac GitHub Action )

PR #346 removes GNU Compiler Collection 12 and replaces it with version 15
in the "Mac" GitHub Action, which tries to build HEMCO on MacOS.  GCC 12
evidently has been "delisted" from the Mac "HomeBrew" package manager,
and this causes the compilation test with GCC 12 to fail.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>